+++ src/diagnostics/unknown_classes.rs
@@ -61,7 +61,7 @@ impl Backend {
 
         let file_use_map: HashMap<String, String> = self.file_use_map(uri);
 
-        let file_namespace: Option<String> = self.namespace_map.read().get(uri).cloned().flatten();
+        let file_namespace: Option<String> = self.first_file_namespace(uri);
 
         let local_classes: Vec<ClassInfo> = self
             .ast_map